Now pull message handlers and handshakes out of AbstractServer
`AbstractServer` now only keeps track of attached `ServerObserver` and notifies them whenever a connection is
established. It is up to the observer (`MapToolServerConnection`) to attach message handlers and perform the
handshake. This avoids the need for `HandshakeProvider<>`, and means `Handshake<>` is no longer in `clientserver`.
