+++ src/main/java/net/rptools/maptool/client/MapToolConnection.java
@@ -19,11 +19,11 @@ import java.util.ArrayList;
 import java.util.List;
 import net.rptools.clientserver.ConnectionFactory;
 import net.rptools.clientserver.simple.DisconnectHandler;
-import net.rptools.clientserver.simple.Handshake;
 import net.rptools.clientserver.simple.connection.Connection;
 import net.rptools.maptool.client.ui.ActivityMonitorPanel;
 import net.rptools.maptool.model.player.LocalPlayer;
 import net.rptools.maptool.server.ClientHandshake;
+import net.rptools.maptool.server.Handshake;
 import net.rptools.maptool.server.ServerConfig;
 import net.rptools.maptool.server.proto.Message;
 import org.apache.logging.log4j.LogManager;
