+++ src/main/java/net/rptools/maptool/server/ClientHandshake.java
@@ -32,7 +32,6 @@ import javax.crypto.BadPaddingException;
 import javax.crypto.IllegalBlockSizeException;
 import javax.crypto.NoSuchPaddingException;
 import javax.swing.*;
-import net.rptools.clientserver.simple.Handshake;
 import net.rptools.clientserver.simple.MessageHandler;
 import net.rptools.clientserver.simple.connection.Connection;
 import net.rptools.lib.MD5Key;
@@ -217,7 +216,7 @@ public class ClientHandshake implements Handshake<Void>, MessageHandler {
       }
 
     } catch (Exception e) {
-      log.warn(e.toString());
+      log.warn("Unexpected exception during client handshake", e);
       setCurrentState(State.Error);
       future.completeExceptionally(new Failure("Handshake.msg.unexpectedError", e));
     }
